requirements.txt
ccxt==4.3.86
pandas==2.2.2
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
